feat: se actualiza la tabla de producthasorder

+++ petcalli_backend 2/src/main/java/mx/petcalli/app/controller/ProductHasOrderController.java
@@ -7,7 +7,7 @@ import mx.petcalli.app.model.ProductHasOrder;
 public interface ProductHasOrderController {
 	ResponseEntity<ProductHasOrder> createProductHasOrder(ProductHasOrder productHasOrder);
 	ResponseEntity< Iterable<ProductHasOrder>> getAllProductHasOrder();
-	ResponseEntity< ProductHasOrder> getProductHasOrderByCompositeId(Long productId, Long orderId);
+	ResponseEntity< ProductHasOrder> getProductHasOrderByCompositeId(Integer productId, Integer orderId);
 	ResponseEntity< Iterable<ProductHasOrder>> getProductHasOrderByOrderId(Integer orderId);
 	ResponseEntity<ProductHasOrder> updateProductHasOrder(ProductHasOrder productHasOrder, Integer productId, Integer orderId);
 	ResponseEntity< Void > deleteProductHasOrderByCompositeId(Integer orderId, Integer productId);
